Teleporting script
Lets you move the character to another location so that we can teleport to different floors of the learning factory. 
To use script drag it onto an object. Then for "Teleport Target" set it as another object you want the character to teleport to. 
For "The Player" set it as PlayerCapsule. 
Now when you walk into the object, it will teleport your player to the other object. 
Note: Also go into Edit->Project Settings->Physics-> MAKE SURE Auto Sync Transforms is checked.